+++ src/utils/execute.go
@@ -4,10 +4,12 @@ import (
 	"bufio"
 	"bytes"
 	"fmt"
+	"github.com/go-cmd/cmd"
 	"io"
 	"log"
 	"os"
 	"os/exec"
+	"time"
 )
 
 // HandleIoError : Handle errors for functions in this file
@@ -39,6 +41,35 @@ func copyAndCapture(w io.Writer, r io.Reader) ([]byte, error) {
 			return out, err
 		}
 	}
+}
+
+// ExecuteStream : run a bash command and stream output
+func ExecuteStream(command string, cwd string, fn func(stdout string, stderr string)) (commandExe *cmd.Cmd) {
+	cmdOptions := cmd.Options{
+		Buffered:  false,
+		Streaming: true,
+	}
+
+	exeCmd := cmd.NewCmdOptions(cmdOptions, "bash", "-c", command)
+
+	go func() {
+		for {
+			select {
+			case line := <-exeCmd.Stdout:
+				fn(line, "")
+			case line := <-exeCmd.Stderr:
+				fn("", line)
+			}
+		}
+	}()
+
+	<-exeCmd.Start()
+
+	for len(exeCmd.Stdout) > 0 || len(exeCmd.Stderr) > 0 {
+		time.Sleep(10 * time.Millisecond)
+	}
+
+	return exeCmd
 }
 
 // Execute : run a bash command