+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added linting guideline to readme for contributors
 - Remove old containers on stacks rm
 
+### Fixed
+- Logs output now includes stderr
+
 ### Changed
 - Changed the order of the readme file
 - Lint with gofmt